refactor: migrate auth utilities to ts
