+++ src/auth.ts
@@ -27,3 +27,4 @@ const result = CreateAuth({
 	},
 });
 export const { auth } = result;
+
