--- src/client-integrations/chakra-ui.jsx
@@ -1,3 +1,0 @@
-'use client';
-
-export * from '@chakra-ui/react';
