--- src/client-integrations/recharts.jsx
@@ -1,3 +1,0 @@
-'use client';
-
-export * from 'recharts';
